Experiment: Speed Up DMCK path execution.
This patch tries to speed up DMCK path execution time by preparing the DMCK
to be able to set the steadyStateTimeout to 0 ms. A side note is that,
if the DMCK steadyStateTimeout=0, the waitEndExploration needs to be set > 100ms,
so that DMCK still exercise all events instead of terminating in the middle of
running the Cassandra-6023 guided path. The DMCK might be terminated earlier because
it does not see any more event in the queue. Another note is that the DMCK might
fail to still reproduce the Cassandra-6023 with steadyStateTimeout sets to 0 due to
the Cassandra node execution time might race to one another, causing them to not
follow the DMCK order of executions.

This patch also refine the nits of the whole project code. From now, we will use
spaces to replace tabs.