Bug fixed the Quick Event Release Algorithm
This patch focused on fixing the Quick Event Release Algorithm.
Previous Algorithm has some flaws in the order of determining the new causal events of an executed event,
determining the local state prediction, and determining whether an event should be
quickly released or not.

Otherwise, this patch also add more logs and remove obsolete logs.